reorganized test suite
