+++ test/unit_tests/test_compiler/foldingtest.c
