+++ test/unit_tests/test_compiler/literaltest.c
