+++ test/unit_tests/test_methods/strtest.c
